Finish course adding/updating. Add pages for error's description

+++ src/main/java/org/example/edutrackerteach/controller/CourseController.java
@@ -1,20 +1,27 @@
 package org.example.edutrackerteach.controller;
 
+import io.minio.errors.*;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.edutrackerteach.dto.course.CourseRequestAdd;
 import org.example.edutrackerteach.dto.course.CourseResponseViewAll;
 import org.example.edutrackerteach.entity.Course;
+import org.example.edutrackerteach.entity.UserDetailsImpl;
 import org.example.edutrackerteach.service.CourseService;
 import org.example.edutrackerteach.service.LessonService;
 import org.example.edutrackerteach.service.TaskService;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.io.IOException;
+import java.security.InvalidKeyException;
+import java.security.NoSuchAlgorithmException;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -34,24 +41,27 @@ public class CourseController {
         return new ModelAndView("course/edit");
     }
     @GetMapping("/edit/{id}")
-    public ModelAndView edit(@PathVariable Integer id) {
+    public ModelAndView edit(@PathVariable Integer id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if(!courseService.isCourseAssignedToProfessor(userDetails.getProfessor().getId(), id))throw new AccessDeniedException("You don't have access to this page");
         return new ModelAndView("course/edit", "course", courseService.getById(id));
     }
     @GetMapping("/get-all")
-    public ResponseEntity<Page<CourseResponseViewAll>> getAll(@RequestParam int page, @RequestParam int pageSize){
-        return ResponseEntity.ok(courseService.getAll(page, pageSize));
+    public ResponseEntity<Page<CourseResponseViewAll>> getAll(@RequestParam int page, @RequestParam int pageSize, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        return ResponseEntity.ok(courseService.getAll(page, pageSize, userDetails.getProfessor().getId()));
     }
     @PostMapping("/add")
-    public ResponseEntity<?> add(@ModelAttribute @Valid CourseRequestAdd courseRequestAdd) {
-        return ResponseEntity.ok(courseService.add(courseRequestAdd));
+    public ResponseEntity<?> add(@ModelAttribute @Valid CourseRequestAdd courseRequestAdd, @AuthenticationPrincipal UserDetailsImpl userDetails) throws ServerException, InsufficientDataException, ErrorResponseException, IOException, NoSuchAlgorithmException, InvalidKeyException, InvalidResponseException, XmlParserException, InternalException {
+        return ResponseEntity.ok(courseService.add(courseRequestAdd, userDetails.getProfessor()));
     }
     @DeleteMapping("/remove")
-    public ResponseEntity<String> remove(@RequestParam long id) {
+    public ResponseEntity<String> remove(@RequestParam long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if(!courseService.isCourseAssignedToProfessor(userDetails.getProfessor().getId(), id))throw new AccessDeniedException("You don't have access to this page");
         courseService.removeById(id);
         return ResponseEntity.ok("removed");
     }
     @GetMapping("/statistic")
-    public ResponseEntity<Map<String, String>> statistic(@RequestParam long id){
+    public ResponseEntity<Map<String, String>> statistic(@RequestParam long id, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        if(!courseService.isCourseAssignedToProfessor(userDetails.getProfessor().getId(), id))throw new AccessDeniedException("You don't have access to this page");
         Map<String, String> statistic = new HashMap<>();
         Course course = courseService.getById(id);
